Reorganized import routines into a folder
